+++ example/src/Screens/TabView/CustomTabBar.tsx
@@ -51,7 +51,7 @@ const CustomTabBar = () => {
       position,
     }: {
       navigationState: State;
-      position: Animated.AnimatedInterpolation;
+      position: Animated.AnimatedInterpolation<number>;
     }) =>
     ({ route, index }: { route: Route; index: number }) => {
       const inputRange = navigationState.routes.map((_, i) => i);