+++ packages/core/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [6.4.9](https://github.com/react-navigation/react-navigation/compare/@react-navigation/core@6.4.8...@react-navigation/core@6.4.9) (2023-06-22)
+
+**Note:** Version bump only for package @react-navigation/core
+
 ## [6.4.8](https://github.com/react-navigation/react-navigation/compare/@react-navigation/core@6.4.7...@react-navigation/core@6.4.8) (2023-02-26)
 
 ### Bug Fixes