+++ packages/native/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [6.1.7](https://github.com/react-navigation/react-navigation/compare/@react-navigation/native@6.1.6...@react-navigation/native@6.1.7) (2023-06-22)
+
+**Note:** Version bump only for package @react-navigation/native
+
 ## [6.1.6](https://github.com/react-navigation/react-navigation/compare/@react-navigation/native@6.1.5...@react-navigation/native@6.1.6) (2023-02-26)
 
 ### Bug Fixes