+++ packages/stack/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [6.3.17](https://github.com/react-navigation/react-navigation/compare/@react-navigation/stack@6.3.16...@react-navigation/stack@6.3.17) (2023-06-22)
+
+**Note:** Version bump only for package @react-navigation/stack
+
 ## [6.3.16](https://github.com/react-navigation/react-navigation/compare/@react-navigation/stack@6.3.15...@react-navigation/stack@6.3.16) (2023-02-26)
 
 ### Bug Fixes